+++ wazo_auth/plugins/backends/tests/test_ldap_user.py
@@ -367,7 +367,7 @@ class TestVerifyPassword(BaseTestCase):
         )
         wazo_ldap.perform_bind.assert_called_once_with(self.expected_user_dn, 'bar')
 
-    def test_that_verify_password_works_using_hostname(self, wazo_ldap):
+    def test_that_verify_password_works_using_domain_name(self, wazo_ldap):
         backend = LDAPUser()
         backend.load(
             {
@@ -383,13 +383,13 @@ class TestVerifyPassword(BaseTestCase):
         wazo_ldap.perform_bind.return_value = True
         wazo_ldap.perform_search.return_value = self.search_obj_result
         self.list_users.return_value = [{'uuid': 'alice-uuid'}]
-        args = {'hostname': 'wazo.io'}
+        args = {'domain_name': 'wazo.io'}
 
         result = backend.verify_password('foo', 'bar', args)
 
         assert_that(result, equal_to(True))
 
-    def test_that_verify_password_using_non_existing_hostname_returns_false(
+    def test_that_verify_password_using_non_existing_domain_name_returns_false(
         self, wazo_ldap
     ):
         backend = LDAPUser()
@@ -408,7 +408,7 @@ class TestVerifyPassword(BaseTestCase):
         wazo_ldap.perform_search.return_value = self.search_obj_result
         self.list_users.return_value = [{'uuid': 'alice-uuid'}]
         self.tenant_service.list_.return_value = []
-        args = {'hostname': 'gmail.com'}
+        args = {'domain_name': 'gmail.com'}
 
         result = backend.verify_password('foo', 'bar', args)