+++ wazo_auth/plugins/http/tokens/tests/test_schemas.py
@@ -67,15 +67,15 @@ class TestTokenRequestSchema(TestCase):
             raises(ValidationError).matching(has_properties(field_name='_schema')),
         )
 
-    def test_that_ldap_backend_using_both_tenant_id_and_hostname_raises_400(self):
-        body = {'backend': 'ldap_user', 'tenant_id': 'x', 'hostname': 'wazo.io'}
+    def test_that_ldap_backend_using_both_tenant_id_and_domain_name_raises_400(self):
+        body = {'backend': 'ldap_user', 'tenant_id': 'x', 'domain_name': 'wazo.io'}
 
         assert_that(
             calling(self.schema.load).with_args(body),
             raises(ValidationError).matching(has_properties(field_name='_schema')),
         )
 
-    def test_that_ldap_backend_requires_tenant_id_or_hostname(self):
+    def test_that_ldap_backend_requires_tenant_id_or_domain_name(self):
         body = {'backend': 'ldap_user'}
 
         assert_that(
@@ -84,7 +84,7 @@ class TestTokenRequestSchema(TestCase):
         )
 
         assert_that(
-            calling(self.schema.load).with_args({'hostname': 'x', **body}),
+            calling(self.schema.load).with_args({'domain_name': 'x', **body}),
             not_(raises(Exception)),
         )